requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
trakt.py==4.4.0
python-dateutil==2.8.2
requests==2.31.0
APScheduler==3.10.4
